+++ rust/types/src/rtp_parameters.rs
@@ -575,9 +575,6 @@ pub enum RtpHeaderExtensionUri {
     /// urn:3gpp:video-orientation
     #[serde(rename = "urn:3gpp:video-orientation")]
     VideoOrientation,
-    /// urn:ietf:params:rtp-hdrext:toffset
-    #[serde(rename = "urn:ietf:params:rtp-hdrext:toffset")]
-    TimeOffset,
     /// <http://www.webrtc.org/experiments/rtp-hdrext/abs-capture-time>
     #[serde(rename = "http://www.webrtc.org/experiments/rtp-hdrext/abs-capture-time")]
     AbsCaptureTime,
@@ -608,7 +605,6 @@ impl FromStr for RtpHeaderExtensionUri {
             "urn:ietf:params:rtp-hdrext:ssrc-audio-level" => Ok(Self::SsrcAudioLevel),
             "https://aomediacodec.github.io/av1-rtp-spec/#dependency-descriptor-rtp-header-extension" => Ok(Self::DependencyDescriptor),
             "urn:3gpp:video-orientation" => Ok(Self::VideoOrientation),
-            "urn:ietf:params:rtp-hdrext:toffset" => Ok(Self::TimeOffset),
             "http://www.webrtc.org/experiments/rtp-hdrext/abs-capture-time" => {
                 Ok(Self::AbsCaptureTime)
             }
@@ -638,7 +634,6 @@ impl RtpHeaderExtensionUri {
             RtpHeaderExtensionUri::SsrcAudioLevel => "urn:ietf:params:rtp-hdrext:ssrc-audio-level",
             RtpHeaderExtensionUri::DependencyDescriptor => "https://aomediacodec.github.io/av1-rtp-spec/#dependency-descriptor-rtp-header-extension",
             RtpHeaderExtensionUri::VideoOrientation => "urn:3gpp:video-orientation",
-            RtpHeaderExtensionUri::TimeOffset => "urn:ietf:params:rtp-hdrext:toffset",
             RtpHeaderExtensionUri::AbsCaptureTime => {
                 "http://www.webrtc.org/experiments/rtp-hdrext/abs-capture-time"
             }