+++ worker/include/RTC/RtpDictionaries.hpp
@@ -123,10 +123,9 @@ namespace RTC
 			SSRC_AUDIO_LEVEL       = 6,
 			DEPENDENCY_DESCRIPTOR  = 7,
 			VIDEO_ORIENTATION      = 8,
-			TIME_OFFSET            = 9,
-			ABS_CAPTURE_TIME       = 10,
-			PLAYOUT_DELAY          = 11,
-			MEDIASOUP_PACKET_ID    = 12
+			ABS_CAPTURE_TIME       = 9,
+			PLAYOUT_DELAY          = 10,
+			MEDIASOUP_PACKET_ID    = 11
 		};
 
 	public: